requirements.txt
# Core Flask
Flask==3.0.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
Werkzeug==3.1.4
gunicorn==21.2.0

# Database
sqlalchemy
psycopg2-binary

# AI
anthropic>=0.18.0

# PDF Generation & Parsing
fpdf2
reportlab
# weasyprint  # Disabled for Railway - requires system deps, using reportlab instead
pdfplumber
pypdf

# Document Processing
python-docx
beautifulsoup4>=4.12.0

# Security & Auth
pyjwt
cryptography
bleach>=6.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# External Services
stripe
twilio
# sendgrid - removed, using Gmail SMTP instead
paramiko

# Utilities
requests
python-dateutil
python-dotenv>=1.0.0

# Testing
locust>=2.20.0
playwright
psutil>=5.9.0
sentry-sdk[flask]>=2.0.0
alembic>=1.13.0
flasgger>=0.9.7
tenacity>=8.2.0
pytz
pre-commit>=4.0.0
